swell-foop: prevent keyboard focus from going to the toolbar Before the keyevents were comming from the clutter stage. Pressing up-arrow would move focus to the toolbar's new-game button. This commit moves the keyevents to the main window.
The game will now always react to keypresses regardless of wether
the stage has focus.

Note: The space-key is reserved for playing the game. If the toolbar
has focus we need to prevent accidentially pressing the new-game
button.

The return-key is still free to allow users of the keyboard to
activate button.