meson: Change config.h handling
The `config.h` file handling, which is done through the
`config.vapi` file, has been changed following Vala
recommendations[0].

To do this, the global variables in the `config.vapi` file has been
moved to the `Config` namespace.

The `DATADIR` and `LOCALEDIR` variables, although they still exist
in the `config.vapi` file, they are not set in the `config.h` file
because they will be set by the compiler's macro definitions when
using compiler flags.

https://bugzilla.gnome.org/show_bug.cgi?id=793340

[0] https://wiki.gnome.org/Projects/Vala/GameDevelopmentSeries/Setup